+++ crates/diagnostics/src/manager.rs
@@ -16,6 +16,10 @@ pub struct Manager {
 impl Manager {
     /// Updates the syntax-based diagnostics for the given document.
     pub fn update_syntax(&mut self, workspace: &Workspace, document: &Document) {
+        if !Self::is_relevant(document) {
+            return;
+        }
+
         self.grammar.remove(&document.uri);
         super::grammar::tex::update(document, workspace.config(), &mut self.grammar);
         super::grammar::bib::update(document, &mut self.grammar);
@@ -49,7 +53,10 @@ impl Manager {
             }
         }
 
-        for document in workspace.iter() {
+        for document in workspace
+            .iter()
+            .filter(|document| Self::is_relevant(document))
+        {
             let project = workspace.project(document);
             super::citations::detect_undefined_citations(&project, document, &mut results);
             super::citations::detect_unused_entries(&project, document, &mut results);
@@ -60,6 +67,13 @@ impl Manager {
         super::labels::detect_undefined_and_unused_labels(workspace, &mut results);
 
         let config = &workspace.config().diagnostics;
+
+        results.retain(|uri, _| {
+            workspace
+                .lookup(uri)
+                .map_or(false, |document| Self::is_relevant(document))
+        });
+
         for (_, diagnostics) in &mut results {
             diagnostics.retain(|diagnostic| {
                 filter_regex_patterns(
@@ -71,5 +85,13 @@ impl Manager {
         }
 
         results
+    }
+
+    fn is_relevant(document: &Document) -> bool {
+        match document.owner {
+            Owner::Client => true,
+            Owner::Server => true,
+            Owner::Distro => false,
+        }
     }
 }